+++ Icon/Icon.js
@@ -71,6 +71,15 @@ const IconBase = kind({
 		flip: PropTypes.oneOf(['auto', 'both', 'horizontal', 'vertical']),
 
 		/**
+		 * The current locale as a
+		 * {@link https://tools.ietf.org/html/rfc5646|BCP 47 language tag}.
+		 *
+		 * @type {String}
+		 * @private
+		*/
+		locale: PropTypes.string,
+
+		/**
 		 * Indicates the content's text direction is right-to-left.
 		 *
 		 * This is set automatically when using {@link ui/Icon.Icon}.
@@ -111,8 +120,11 @@ const IconBase = kind({
 		className: ({size, styler}) => styler.append(
 			(typeof size === 'string' ? size : null)
 		),
-		flip: ({flip, rtl}) => {
+		flip: ({children, flip, locale, rtl}) => {
 			if (flip === 'auto') {
+				if (locale === 'he-IL' && children === 'help') {
+					return null;
+				}
 				return rtl ? 'horizontal' : null;
 			}
 
@@ -125,6 +137,7 @@ const IconBase = kind({
 	},
 
 	render: ({css, size, ...rest}) => {
+		delete rest.locale;
 		delete rest.rtl;
 
 		return UiIconBase.inline({
@@ -416,7 +429,7 @@ const IconBase = kind({
 const IconDecorator = compose(
 	Pure,
 	Skinnable,
-	I18nContextDecorator({rtlProp: 'rtl'})
+	I18nContextDecorator({localeProp: 'locale', rtlProp: 'rtl'})
 );
 
 /**